changed integer identifiers, like id and version to Long. It seems like hibernate cannot deduce the proper type to be used from the entity. Instead it tries to figure out the type somehow. This does not always work. E.g. instead of Integer it loads the vale as Double

+++ mongodb/src/test/java/eu/drus/jpa/unit/test/model/Account.java
@@ -28,11 +28,11 @@ public abstract class Account {
 
     @Id
     @GeneratedValue(strategy = GenerationType.TABLE)
-    private Integer id;
+    private Long id;
 
     @Version
     @Column(name = "VERSION")
-    private Integer version;
+    private Long version;
 
     @ManyToOne(optional = false, cascade = {
             CascadeType.DETACH, CascadeType.MERGE, CascadeType.PERSIST, CascadeType.REFRESH
